+++ springboot/src/main/java/com/asakatu/Greeting.java
@@ -1,0 +1,20 @@
+package com.asakatu;
+
+public class Greeting {
+
+    private final long id;
+    private final String content;
+
+    public Greeting(long id, String content) {
+        this.id = id;
+        this.content = content;
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public String getContent() {
+        return content;
+    }
+}
