+++ springboot/src/main/java/com/asakatu/exeptions/UserAlreadyExistException.java
@@ -1,0 +1,12 @@
+package com.asakatu.exeptions;
+
+public class UserAlreadyExistException extends IllegalArgumentException {
+
+    public UserAlreadyExistException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public UserAlreadyExistException(String message) {
+        super(message);
+    }
+}
