+++ springboot/src/main/java/com/asakatu/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.asakatu.repository;
+
+import com.asakatu.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByUsername(String username);
+}
