+++ springboot/src/main/java/com/asakatu/service/UserRegistrationService.java
@@ -1,0 +1,21 @@
+package com.asakatu.service;
+
+import com.asakatu.entity.User;
+import com.asakatu.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserRegistrationService {
+
+    @Autowired
+    private UserRepository userRepository;
+    @Autowired
+    private BCryptPasswordEncoder bCryptPasswordEncoder;
+
+    public void save(User user) {
+        user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
+        userRepository.save(user);
+    }
+}
