--- springboot/src/main/java/hello/WebSecurityConfig.java
@@ -1,164 +1,0 @@
-package hello;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.security.web.WebAttributes;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.security.web.csrf.CookieCsrfTokenRepository;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-import java.io.IOException;
-import java.util.Arrays;
-
-@Configuration
-@EnableWebSecurity
-public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http
-            .authorizeRequests()
-                .mvcMatchers("/greeting", "/user_registration" )
-                    .permitAll()
-                .mvcMatchers( "/")
-                    .hasRole("USER")
-                .anyRequest()
-                    .authenticated()
-            .and()
-                .exceptionHandling()
-                    .authenticationEntryPoint(new SimpleAuthenticationEntryPoint())
-                    .accessDeniedHandler(new SimpleAccessDeniedHandler())
-            .and()
-                // LOGIN
-                .formLogin()
-                    .loginProcessingUrl("/login").permitAll()
-                        .usernameParameter("username")
-                        .passwordParameter("password")
-                    .successHandler(new SimpleAuthenticationSuccessHandler())
-                    .failureHandler(new SimpleAuthenticationFailureHandler())
-            .and()
-                .cors()
-            .and()
-                .csrf()
-                    .disable()
-        ;
-    }
-
-
-    @Bean
-    CorsConfigurationSource corsConfigurationSource() {
-        CorsConfiguration corsConfiguration = new CorsConfiguration();
-        corsConfiguration.setAllowedMethods(Arrays.asList("GET","POST","PUT","DELETE","OPTIONS"));
-        corsConfiguration.addAllowedOrigin("http://localhost:9000");
-        corsConfiguration.setAllowCredentials(true);
-
-        UrlBasedCorsConfigurationSource corsSource = new UrlBasedCorsConfigurationSource();
-        corsSource.registerCorsConfiguration("/**", corsConfiguration);
-        return corsSource;
-    }
-
-
-    @Autowired
-    public void configureGlobal(AuthenticationManagerBuilder auth,
-                                @Qualifier("simpleUserDetailsService") UserDetailsService userDetailsService,
-                                PasswordEncoder passwordEncoder) throws Exception {
-        auth.eraseCredentials(true)
-                .userDetailsService(userDetailsService)
-                .passwordEncoder(passwordEncoder);
-    }
-
-
-    @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-}
-
-class SimpleAuthenticationSuccessHandler implements AuthenticationSuccessHandler {
-
-    @Override
-    public void onAuthenticationSuccess(HttpServletRequest request,
-                                        HttpServletResponse response,
-                                        Authentication auth) throws IOException, ServletException {
-        if (response.isCommitted()) {
-//            log.info("Response has already been committed.");
-            return;
-        }
-
-        response.setStatus(HttpStatus.OK.value());
-        clearAuthenticationAttributes(request);
-    }
-
-    /**
-     * Removes temporary authentication-related data which may have been stored in the
-     * session during the authentication process.
-     */
-    private void clearAuthenticationAttributes(HttpServletRequest request) {
-        HttpSession session = request.getSession(false);
-
-        if (session == null) {
-            return;
-        }
-        session.removeAttribute(WebAttributes.AUTHENTICATION_EXCEPTION);
-    }
-}
-
-class SimpleAuthenticationFailureHandler implements AuthenticationFailureHandler {
-
-    @Override
-    public void onAuthenticationFailure(HttpServletRequest request,
-                                        HttpServletResponse response,
-                                        AuthenticationException exception) throws IOException, ServletException {
-        response.sendError(HttpStatus.FORBIDDEN.value(), HttpStatus.FORBIDDEN.getReasonPhrase());
-    }
-
-}
-
-
-class SimpleAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-    @Override
-    public void commence(HttpServletRequest request,
-                         HttpServletResponse response,
-                         AuthenticationException exception) throws IOException, ServletException {
-        if (response.isCommitted()) {
-//            log.info("Response has already been committed.");
-            return;
-        }
-        response.sendError(HttpStatus.UNAUTHORIZED.value(), HttpStatus.UNAUTHORIZED.getReasonPhrase());
-    }
-
-}
-
-class SimpleAccessDeniedHandler implements AccessDeniedHandler {
-
-    @Override
-    public void handle(HttpServletRequest request,
-                       HttpServletResponse response,
-                       AccessDeniedException exception) throws IOException, ServletException {
-        response.sendError(HttpStatus.FORBIDDEN.value(), HttpStatus.FORBIDDEN.getReasonPhrase());
-    }
-
-}
