--- springboot/src/main/java/hello/exeptions/UserAlreadyExistException.java
@@ -1,12 +1,0 @@
-package hello.exeptions;
-
-public class UserAlreadyExistException extends IllegalArgumentException {
-
-    public UserAlreadyExistException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public UserAlreadyExistException(String message) {
-        super(message);
-    }
-}
