--- springboot/src/main/java/hello/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package hello.repository;
-
-import hello.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByUsername(String username);
-}
